Sensor names rework, optimization

+++ alienfan-tools/alienfan-shared/ConfigFan.h
@@ -55,7 +55,7 @@ public:
 
 	vector<fan_overboost> boosts;
 	map<byte, string> powers;
-	map<byte, string> sensors;
+	map<WORD, string> sensors;
 
 	ConfigFan();
 	~ConfigFan();
